Simplify fcm custom service and remove manifest check
