+++ AndroidSDKFcm/src/main/java/com/leanplum/LeanplumPushFirebaseMessagingService.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016, Leanplum, Inc. All rights reserved.
+ * Copyright 2020, Leanplum, Inc. All rights reserved.
  *
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
@@ -21,41 +21,29 @@
 
 package com.leanplum;
 
-import android.annotation.SuppressLint;
-import android.os.Build;
-import android.os.Bundle;
-
+import androidx.annotation.NonNull;
 import com.google.firebase.messaging.FirebaseMessagingService;
 import com.google.firebase.messaging.RemoteMessage;
-import com.leanplum.internal.Constants;
-import com.leanplum.internal.Log;
-import com.leanplum.internal.Util;
-
-import java.util.Map;
 
 /**
  * FCM listener service, which enables handling messages on the app's behalf.
  *
  * @author Anna Orlova
  */
-@SuppressLint("Registered")
 public class LeanplumPushFirebaseMessagingService extends FirebaseMessagingService {
+
+  private final LeanplumFirebaseServiceHandler handler = new LeanplumFirebaseServiceHandler();
 
   @Override
   public void onCreate() {
     super.onCreate();
-    Leanplum.setApplicationContext(getApplicationContext());
+    handler.onCreate(getApplicationContext());
   }
 
   @Override
-  public void onNewToken(String token) {
+  public void onNewToken(@NonNull String token) {
     super.onNewToken(token);
-
-    LeanplumPushService.setCloudMessagingProvider(new LeanplumFcmProvider());
-    LeanplumPushService.getCloudMessagingProvider().storePreferences(getApplicationContext(), token);
-
-    //send the new token to backend
-    LeanplumPushService.getCloudMessagingProvider().onRegistrationIdReceived(getApplicationContext(), token);
+    handler.onNewToken(token, getApplicationContext());
   }
 
   /**
@@ -65,28 +53,7 @@ public class LeanplumPushFirebaseMessagingService extends FirebaseMessagingServi
    * @param remoteMessage Object representing the message received from Firebase Cloud Messaging.
    */
   @Override
-  public void onMessageReceived(RemoteMessage remoteMessage) {
-    try {
-      Map<String, String> messageMap = remoteMessage.getData();
-      if (messageMap.containsKey(Constants.Keys.PUSH_MESSAGE_TEXT)) {
-        LeanplumPushService.handleNotification(getApplicationContext(), getBundle(messageMap));
-      }
-      Log.d("Received push notification message: %s", messageMap.toString());
-    } catch (Throwable t) {
-      Log.exception(t);
-    }
-  }
-
-  /**
-   * @param messageMap {@link RemoteMessage}'s data map.
-   */
-  private Bundle getBundle(Map<String, String> messageMap) {
-    Bundle bundle = new Bundle();
-    if (messageMap != null) {
-      for (Map.Entry<String, String> entry : messageMap.entrySet()) {
-        bundle.putString(entry.getKey(), entry.getValue());
-      }
-    }
-    return bundle;
+  public void onMessageReceived(@NonNull RemoteMessage remoteMessage) {
+    handler.onMessageReceived(remoteMessage, getApplicationContext());
   }
 }
